Cancel drag and drop when pointers are canceled for the drag window
When a cancel event is synthesized for any reason for the drag window
when drag and drop is active, we should cancel the ongoing drag and drop
operation. Otherwise, the system may end up in an unexpected state
resulting in unwanted behavior.

Bug: 291181957
Test: atest inputflinger_tests

Change-Id: Ic5e328f0e445e3271cb1797725c4c00589a6b5e1
Merged-In: Ic5e328f0e445e3271cb1797725c4c00589a6b5e1
Signed-off-by: Linnan Li <lilinnan@xiaomi.corp-partner.google.com>